static constructors and builders for Spec interfaces
