+++ simplesaga-action/src/main/java/io/simplesource/saga/action/async/AsyncSpec.java
@@ -2,12 +2,9 @@ package io.simplesource.saga.action.async;
 
 import io.simplesource.data.Result;
 import io.simplesource.saga.model.messages.UndoCommand;
-import io.simplesource.saga.model.serdes.TopicSerdes;
-import io.simplesource.saga.shared.topics.TopicCreation;
 import lombok.*;
 
 import java.time.Duration;
-import java.util.List;
 import java.util.Optional;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
@@ -19,22 +16,10 @@ import java.util.function.Function;
   * @param <O> - value returned by async function
   * @param <R> - final result type that ends up in value topic
   */
-@Value(staticConstructor = "of")
+@Value
 @Builder
-@AllArgsConstructor
+@AllArgsConstructor(staticName = "of")
 public final class AsyncSpec<A, D, K, O, R> {
-
-    public interface UndoFunction<A, D, K, R> {
-        Optional<UndoCommand<A>> apply(D decodedInput, K outputKey, R result);
-    }
-
-    @Value(staticConstructor = "of")
-    public static final class AsyncResult<A, D, K, O, R> {
-        public final Function<O, Optional<Result<Throwable, R>>> outputMapper;
-        public final Function<D, K> keyMapper;
-        public final UndoFunction<A, D, K, R> undoFunction;
-        public final Optional<TopicSerdes<K, R>> outputSerdes;
-    }
 
     public final String actionType;
     public final Function<A, Result<Throwable, D>> inputDecoder;
