+++ simplesaga-http/src/main/java/io/simplesource/saga/action/http/HttpSpec.java
@@ -1,6 +1,8 @@
 package io.simplesource.saga.action.http;
 
 import io.simplesource.saga.action.async.Callback;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Value;
 
 import java.time.Duration;
@@ -14,7 +16,10 @@ import java.util.function.BiConsumer;
   * @param <O> - value returned by the Http request - also normally quite generic
   * @param <R> - final result type that ends up in value topic
   */
+
 @Value
+@Builder
+@AllArgsConstructor(staticName = "of")
 public final class HttpSpec<A, K, B, O, R> {
     public final String actionType;
     public final HttpRequest.HttpRequestDecoder<A, K, B> decoder;
